Java - Graphics
This pong game has been programme as a 2 player game when the players can move the side panels set to the same speed and a ball that hits the panels, if a player misses the ball the opposing player will gain a point. Once a player reaches 3 points the winner in declared.

This game has been created with once class I called PongG and has 3 booleans for each mode a title screen for the player to enter the game, playing mode where the game is played and game over where the winning player is displayed and the game terminates. 

The code is this file can easily be manipulated to create a longer lasting game that last longer that 3 points